Add validation for /api/token and fix docs
